requirements.txt
faker==24.0.0
mysql-connector-python==8.3.0
